requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Jinja2==3.1.2
Werkzeug==2.2.3
requests==2.31.0
SQLAlchemy==2.0.20
torch==2.1.0
openai-whisper==20230918
numpy>=1.23.5